fir: Split up crate in multiple modules
