+++ fir/src/iter/multi_mapper.rs
@@ -1,0 +1,1 @@
+pub trait MultiMapper {}
